updated tooltip with latest changes
