+++ src/overview/src/tooltip.js
@@ -1,12 +1,13 @@
 import React from 'react'
-import styled from 'styled-components'
 
 import Container from './container'
 import { Tooltip } from '../../components'
 
 const Tooltips = () => (
   <Container title="Tooltip">
-    <Tooltip>I am a tooltip</Tooltip>
+    <Tooltip>I am a tooltip on top</Tooltip>
+    <br />
+    <Tooltip bottom>I am a tooltip on bottom</Tooltip>
   </Container>
 )
